menambahkan fungsi event dan manage event - main business flow

+++ src/helpers/fileRemover.helper.js
@@ -1,12 +1,18 @@
 const fs = require("fs")
 const fileRemover = (file) => {
-    if(file){
-        const fileName = `uploads/${file.filename}`
-        fs.unlink(fileName, (err)=>{
-            if(err){
-                throw Error(err.message)
+    try {
+        if(file){
+            const fileName = `uploads/${file.filename}`
+            if(fileName){
+                fs.unlink(fileName, (err)=>{
+                    if(err){
+                        return console.log(err)
+                    }
+                })
             }
-        })
+        }
+    } catch (err) {
+        throw Error("error")
     }
 }
